requirements.txt
# Core dependencies
Flask==3.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
PyYAML==6.0.1

# Hardware (Raspberry Pi)
RPi.GPIO==0.7.1
adafruit-circuitpython-pca9685==3.4.7
adafruit-blinka==8.27.0

# OBD2
obd==0.7.1

# Testing
pytest==7.4.3
pytest-html==4.1.1
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code quality
black==23.12.1
ruff==0.1.9

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Telemetry Visualization (charts rendered client-side with Apache ECharts)
pandas==2.1.4
